Revert "Fixing Bad Merge, Recommiting Changes to Datatables and Editing Appointments"
This reverts commit 3cbff9615843788fd67f8d8c7c078a4013ff7d8b.

+++ FacultyAdvisementSite/src/java/FacultyAdvisement/AppointmentBean.java
@@ -130,7 +130,11 @@ public class AppointmentBean implements Serializable {
         return "/customerFolder/signup.xhtml";
     }
     
-    
+      public String editDesired(String key, Appointment appointment) throws IOException, SQLException
+    {   this.appointment = appointment;
+        //this.updateAppointment(key, appointment, false);
+        return "/customerFolder/signup.xhtml";
+    }
     
     public void insertAppointment() throws SQLException {
         Connection conn1 = ds.getConnection();
@@ -360,9 +364,13 @@ public class AppointmentBean implements Serializable {
 
             ps.execute();
             
-            ps = conn.prepareStatement("delete FROM desired where ID =?");
-            ps.setString(1,Long.toString(appointment.aID));
-            ps.execute();
+            if(!isCancel)
+            {
+                 ps = conn.prepareStatement("delete FROM desired where ID =?");
+                 ps.setString(1,Long.toString(appointment.aID));
+                 ps.execute();
+            }
+           
         } finally {
             conn.close();
         }
